requirements.txt
streamlit
pandas
gspread
google-auth
google-api-python-client